+++ example/fetch_homework.dart
@@ -3,7 +3,7 @@ import 'package:lectio_wrapper/lectio_wrapper.dart';
 void main() async {
   Account account = Account(
       0 /* yours schools id */, "" /* your username */, "" /* your password */);
-  Student? student = await account.login();
+  Student? student = await account.login(autologin: false);
   var homework = await student!.homework.list();
   print(homework);
 }