+++ lib/types/terms/term.dart
@@ -1,0 +1,6 @@
+class Term {
+  int term;
+  String name;
+  bool active;
+  Term(this.term, this.name, this.active);
+}