+++ test/assignment/assignment_test.dart
@@ -9,7 +9,7 @@ void main() {
   Account account =
       Account(int.parse(env['GYM_ID']!), env['USERNAME']!, env['PASSWORD']!);
   Student? student;
-  setUp(() async => {student = await account.login()});
+  setUp(() async => {student = await account.login(autologin: false)});
 
   test('list assignments', () async {
     var assignments = await student!.assignments.list(2022);